"Moved mock classes to support/../mocks.py"
